Create FunctionScheduler test service
